requirements.txt
kafka-python==2.0.2
pyyaml==6.0.1
python-dotenv==1.0.0
cryptography==41.0.7 